requirements.txt
# FastAPI 和相關套件
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 資料庫相關
sqlalchemy==2.0.23
alembic==1.12.1

# 驗證和安全
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0

# Pydantic (資料驗證)
pydantic==2.5.0

# LangChain 和 RAG 相關
langchain==0.0.335
langchain-community==0.0.5
langchain-google-genai==0.0.5
google-generativeai==0.3.2

# PDF 處理
pypdf==3.17.0

# 向量資料庫
chromadb==0.4.17

# HTTP 客戶端
httpx==0.25.2

# CORS 支援 (包含在 FastAPI 中，但明確列出)
starlette==0.27.0
